Move test data factory packages.
