+++ src/main/kotlin/com/openlattice/chronicle/authorization/principals/PrincipalMapstore.kt
@@ -25,7 +25,7 @@ import com.openlattice.chronicle.authorization.SecurablePrincipal
 import com.openlattice.chronicle.hazelcast.HazelcastMap
 import com.openlattice.chronicle.postgres.ResultSetAdapters
 import com.openlattice.chronicle.storage.ChroniclePostgresTables.Companion.PRINCIPALS
-import com.openlattice.chronicle.util.TestDataFactory
+import com.openlattice.chronicle.util.tests.TestDataFactory
 import com.geekbeast.postgres.PostgresArrays
 import com.geekbeast.postgres.mapstores.AbstractBasePostgresMapstore
 import com.zaxxer.hikari.HikariDataSource
