+++ src/main/kotlin/com/openlattice/chronicle/mapstores/authorization/PermissionMapstore.kt
@@ -36,7 +36,7 @@ import com.openlattice.chronicle.storage.ChroniclePostgresTables.Companion.SECUR
 import com.openlattice.chronicle.storage.PostgresColumns
 import com.openlattice.chronicle.storage.PostgresColumns.Companion.ACL_KEY
 import com.openlattice.chronicle.storage.PostgresColumns.Companion.SECURABLE_OBJECT_TYPE
-import com.openlattice.chronicle.util.TestDataFactory
+import com.openlattice.chronicle.util.tests.TestDataFactory
 import com.zaxxer.hikari.HikariDataSource
 import org.springframework.stereotype.Component
 import java.sql.Array
