+++ src/main/kotlin/com/openlattice/chronicle/mapstores/storage/StudyMapstore.kt
@@ -14,7 +14,7 @@ import com.openlattice.chronicle.storage.ChroniclePostgresTables.Companion.ORGAN
 import com.openlattice.chronicle.storage.ChroniclePostgresTables.Companion.STUDIES
 import com.openlattice.chronicle.storage.PostgresColumns.Companion.STUDY_ID
 import com.openlattice.chronicle.study.Study
-import com.openlattice.chronicle.util.TestDataFactory
+import com.openlattice.chronicle.util.tests.TestDataFactory
 import com.zaxxer.hikari.HikariDataSource
 import org.springframework.stereotype.Component
 import java.sql.ResultSet
