+++ src/main/kotlin/com/openlattice/chronicle/serializers/AclKeyStreamSerializer.kt
@@ -5,7 +5,7 @@ import com.hazelcast.nio.ObjectDataOutput
 import com.geekbeast.rhizome.hazelcast.serializers.ListStreamSerializers
 import com.openlattice.chronicle.authorization.AclKey
 import com.openlattice.chronicle.hazelcast.StreamSerializerTypeIds
-import com.openlattice.chronicle.util.TestDataFactory
+import com.openlattice.chronicle.util.tests.TestDataFactory
 import com.geekbeast.hazelcast.serializers.TestableSelfRegisteringStreamSerializer
 import org.springframework.stereotype.Component
