+++ src/main/kotlin/com/openlattice/chronicle/serializers/PrincipalSetStreamSerializer.kt
@@ -25,7 +25,7 @@ import com.hazelcast.nio.ObjectDataOutput
 import com.openlattice.chronicle.authorization.Principal
 import com.openlattice.chronicle.authorization.PrincipalSet
 import com.openlattice.chronicle.hazelcast.StreamSerializerTypeIds
-import com.openlattice.chronicle.util.TestDataFactory
+import com.openlattice.chronicle.util.tests.TestDataFactory
 import com.geekbeast.hazelcast.serializers.SetStreamSerializer
 import org.springframework.stereotype.Component
 import java.io.IOException
